requirements.txt
beautifulsoup4==4.12.2
bing_image_downloader==1.1.2
pybase64==1.2.3
Requests==2.31.0
ShopifyAPI==12.3.0
